Add kanboard application with keybinding
So this requires a few things on non-xmonad side. First use a browser
profile, things would be much more controllable that way. Assuming you
use something like Kanboard (only needs to be a url which browser can
visit), when mod+x is hit, it opens up the said url in a new-window and
in kiosk modeᶠ. new-window so that multiple kanboard sessions can be
opened separately in different workspaces, and kiosk mode to make it
full-screen for niceness.

ᶠ Kiosk mode in firefox enables a global variable which then will open
everything in kiosk mode afterwards which may be unwanted, so a better
approach will be to use a separate profile.