Add utility for generating normal and visual mode submaps simulatenously
makeChords accepts a major key and a set of subkeys to generate both the
sets of keys (normal and visual). This will reduce code needed to have
multiple modes for different chords. Selecting either normal mode or
visual mode keys is as simple as doing head and last over the complete
set.