Add support for currently running player for metadata retrieval
Playerctl (by default) chooses the first player for metadata (and
commands). This often can conflict if more than one player are being
played (like spotify in browser and rhythmbox). getRunningPlayer queries
what is the current running player based on status and only retrieves
information from that player. If all player are paused then the first
player is used for metadata.