Remove wins attribute from xmobar config
This was a blockage to first look for adding topbar in floating windows,
but I couldn't find one. Only place where we need a title bar for
floating windows would be to put them in Preferences, but this can be
achieved by shifting the floating window to tiling mode.

Long story short, title bar isn't needed in xmobar to save space.