Revert "Add ewhmDesktopsManageHook for browsers"
This reverts commit 1aa077affd4023439ab51adcc431e0673784dd10.

This is an absolute disaster. The hook completely ruins the kiosk mode
in firefox (which I use for my kanban sessions). It forces the browser
instances to start in tiling mode, and even if it is moved to floating
with a specific size, it doesn't respect that rule after restarting.

In all, an absolute waste of hook and costed me almost an hour to debug
this crap.

DON'T USE.