Add pomodoro plugin
This plugin is a visual aid for xmobar. It runs a timer and puts
appropriate breaks and long breaks. It also provides method to toggle
the state of a pomodoro session.

To interact with the tool, either you can use cmdline or use pomodoro
chords.

  echo "start" > /tmp/pomodoro.ctrl
  echo "skip" > /tmp/pomodoro.ctrl
  echo "toggle" > /tmp/pomodoro.ctrl

Or use the gnome-pomodoro keybindings in pomodoroChords.

You should run both pomodoro's (this and gnome-pomdoro, the keybindings
allow you to run both simulatenously) if you want to store your workflow
in database.

Gnome-pomdoro stores all sprints and breaks in a database, this tool
isn't that vibrant. It only provides a timer on xmobar. So running both
together gives you both things, a visual aid and a database.